Add loading and usage of the pretrained model
Made it possible to save and load the chatbot model so that it can be used in order to make predictions in the CLI. To do this a ChatbotModel class was created and moved from the notebook to a .py file (otherwise the model cannot be loaded from the pretrained state). Also added a wrapper class for the model which loads the pretrained model etc.

The model is now used to predict on the user input when the CLI is running.